force RecreateDb command in development environment

+++ GrayMint.Common.AspNetCore/GrayMintApp.cs
@@ -29,6 +29,10 @@ public static class GrayMintApp
         var appDbContext = scope.ServiceProvider.GetRequiredService<T>();
         if (args.Contains("/recreateDb", StringComparer.OrdinalIgnoreCase))
         {
+            var hostEnvironment = serviceProvider.GetService<IHostEnvironment>();
+            if (hostEnvironment?.IsDevelopment() == true)
+                throw new InvalidOperationException("RecreateDb command is only allowed in Development environment.");
+
             logger.LogInformation($"Recreating the {nameof(T)} database...");
             await appDbContext.Database.EnsureDeletedAsync();
         }
